move TrainingAverager to hivemind.optim
