+++ hivemind/averaging/__init__.py
@@ -1,2 +1,1 @@
 from hivemind.averaging.averager import DecentralizedAverager
-from hivemind.averaging.training import TrainingAverager
